requirements.txt
fastapi
uvicorn
python-dotenv
mangum